+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 ## [Next]
-
+- Add support for JPLS-6H and other JPLS-*H Jupiter device variants
 
 ## [1.3.3] - 2026-01-01
 - Add support for Venus A and Venus D